requirements.txt
streamlit
streamlit-audiorecorder
streamlit-bokeh-events
streamlit-webrtc
streamlit_mic_recorder
webrtcvad
webrtcvad-wheels
websockets
torch
spacy
spacy-legacy
spacy-loggers
SpeechRecognition
python-dateutil
python-docx
python-dotenv
pytz
PyYAML
pyOpenSSL
PyPDF2
numpy
onnxruntime
openai
openai-whisper
packaging
pandas
groq
bokeh==2.4.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastjsonschema